versionCompare(): actually do numerical comparisons
Actually do what 01579703f3b92820fb413e46c10eb5c5526fe70b claimed, re fixing
    "1.0.4" > "1.0.39"

This will break on version numbers containing letters (such as "1.0.39a") but
we plan to use ".99" version numbers for future alpha releases so it should be
fine.

+++ src/tasks/tasks-utils.cpp
@@ -76,12 +76,12 @@ int versionCompare(QString found, QString fixed)
     for(i = 0; i < foundlist.size() - fixedlist.size(); i++)
         fixedlist.append("0");
 
-    /* Compare each portion of the strings. */
+    /* Numerically compare each portion of the strings. */
     for(i = 0; i < fixedlist.size(); i++)
     {
-        if(foundlist[i] > fixedlist[i])
+        if(foundlist[i].toInt() > fixedlist[i].toInt())
             return (1);
-        else if(foundlist[i] < fixedlist[i])
+        else if(foundlist[i].toInt() < fixedlist[i].toInt())
             return (-1);
     }
